backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.1
httpx==0.25.1
python-multipart==0.0.6
aiohttp==3.9.1
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3
